# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.EmployeeListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Employee'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_page:
        type: integer
    type: object
  dto.JobTitleListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.JobTitle'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_page:
        type: integer
    type: object
...
  models.Employee:
    properties:
      department_id:
        type: integer
      email:
        type: string
      employee_code:
        type: string
      first_name:
        type: string
      hire_date:
        type: string
      id:
        type: integer
      job_title_id:
        type: integer
      last_name:
        type: string
      phone:
        type: string
      status:
        type: string
    type: object
  models.JobTitle:
    properties:
      id:
        type: integer
      title:
        type: string
    type: object
...
info:
  contact:
    email: support@hrms.example.com
    name: HRMS Support
    url: http://hrms.example.com/support
  description: Human Resource Management System API built with Go (Gin framework),
    GORM, and MySQL. Supports Employee management, Attendance, Payroll, Leave Requests,
    Departments, Job Titles, and User Authentication/Authorization.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://hrms.example.com/terms/
  title: HRMS API
  version: "1.0"
paths:
...
  /employees:
    get:
      description: Get a paginated list of employees
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EmployeeListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List employees with pagination
      tags:
      - Employees
    post:
      consumes:
      - application/json
      description: Create a new employee with all required fields
      parameters:
      - description: Employee object
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/models.Employee'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Employee'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new employee
      tags:
      - Employees
  /employees/{id}:
    delete:
      description: Delete an employee by ID
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an employee
      tags:
      - Employees
    get:
      description: Get an employee's details by ID
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Employee'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get employee by ID
      tags:
      - Employees
    put:
      consumes:
      - application/json
      description: Update employee details by ID
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Employee object
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/models.Employee'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Employee'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an employee
      tags:
      - Employees
  /job-titles:
    get:
      description: Get a paginated list of job titles
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.JobTitleListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List job titles
      tags:
      - JobTitles
    post:
      consumes:
      - application/json
      description: Create a new job title in the system
      parameters:
      - description: Job Title
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/models.JobTitle'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.JobTitle'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new job title
      tags:
      - JobTitles
  /job-titles/{id}:
    get:
      description: Get a specific job title by ID
      parameters:
      - description: Job Title ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JobTitle'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a job title by ID
      tags:
      - JobTitles
    put:
      consumes:
      - application/json
      description: Update an existing job title by ID
      parameters:
      - description: Job Title ID
        in: path
        name: id
        required: true
        type: integer
      - description: Job Title
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/models.JobTitle'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JobTitle'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a job title
      tags:
      - JobTitles
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by your JWT token. Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"